rust: Add support for ws-protocol time capability
