+++ rust/foxglove/src/websocket.rs
@@ -30,6 +30,8 @@ use tokio_util::sync::CancellationToken;
 mod protocol;
 #[cfg(test)]
 mod tests;
+#[cfg(all(test, feature = "unstable"))]
+mod unstable_tests;
 
 pub(crate) const SUBPROTOCOL: &str = "foxglove.sdk.v1";
 
@@ -465,6 +467,28 @@ impl Server {
                     "Failed to send parameter values to client {}: {err}",
                     client.addr
                 );
+            }
+        }
+    }
+
+    /// Publish the current timestamp to all clients.
+    #[cfg(feature = "unstable")]
+    pub async fn publish_time(&self, timestamp_nanos: u64) {
+        if !self.capabilities.contains(&Capability::Time) {
+            tracing::error!("Server does not support time capability");
+            return;
+        }
+
+        // https://github.com/foxglove/ws-protocol/blob/main/docs/spec.md#time
+        let mut buf = BytesMut::with_capacity(9);
+        buf.put_u8(protocol::server::BinaryOpcode::TimeData as u8);
+        buf.put_u64_le(timestamp_nanos);
+        let message = Message::binary(buf);
+
+        let clients = self.clients.get();
+        for client in clients.iter() {
+            if let Err(err) = client.control_plane_tx.send_async(message.clone()).await {
+                tracing::error!("Failed to send time to client {}: {err}", client.addr);
             }
         }
     }
