+++ rust/foxglove/src/websocket/protocol/server.rs
@@ -14,7 +14,8 @@ use std::collections::HashSet;
 #[repr(u8)]
 pub enum BinaryOpcode {
     MessageData = 1,
-    // TimeData = 2,
+    #[cfg(feature = "unstable")]
+    TimeData = 2,
     // ServiceCallResponse = 3,
     // FetchAssetResponse = 4,
 }
@@ -118,6 +119,13 @@ pub enum Capability {
     /// Allow clients to get & set parameters.
     #[cfg(feature = "unstable")]
     Parameters,
+    /// Inform clients about the latest server time.
+    ///
+    /// This allows accelerated, slowed, or stepped control over the progress of time. If the
+    /// server publishes time data, then timestamps of published messages must originate from the
+    /// same time source.
+    #[cfg(feature = "unstable")]
+    Time,
 }
 
 // https://github.com/foxglove/ws-protocol/blob/main/docs/spec.md#server-info
